Refactor CookieStore -> StatelessSession
